star 1.3

+++ lib/saveargs.c
@@ -1,4 +1,4 @@
-/* @(#)saveargs.c	1.7 98/05/31 Copyright 1995 J. Schilling */
+/* @(#)saveargs.c	1.9 00/05/07 Copyright 1995 J. Schilling */
 /* save argc, argv for command error printing routines */
 /*
  * This program is free software; you can redistribute it and/or modify
@@ -21,6 +21,7 @@
 #include <strdefs.h>
 #include <stdxlib.h>
 #include <avoffset.h>
+#include <schily.h>
 
 #if	!defined(AV_OFFSET) || !defined(FP_INDIR)
 #	ifdef	HAVE_SCANSTACK
@@ -56,7 +57,7 @@ void save_args(ac, av)
 
 	slen = strlen(av[0]) + 1;
 
-	if (slen <= sizeof(av0_sp))
+	if (slen <= (int)sizeof(av0_sp))
 		av0_saved = av0_sp;
 	else
 		av0_saved = malloc(slen);
